Add `jumphost+targethost` support to ssh config
Split `ssh jumphost+targethost` into `ssh jumphost -W targethost:%p`
Primarily useful when the targethost is a generic name that refers to
different servers depending on the jumphost. For example, redis01 is
a different box in datacenter1 vs datacenter2. Implementing this requires
ProxyCommand instead of ProxyJump because ProxyJump doesn't support
splitting the hostname into parts.

Use this in conjunction with a jumphost definition in this file to quickly get onto any host behind any jumphost.