Comment out VisualVM for now
I don't use it these days since I haven't been doing any complicated
Kafka work for several years.

I commented it out rather than removing it completely because in the
past when I wanted this tool, I couldn't remember the name of it. So
leaving in here for now in case I need it in the future. A quick google
serach for `JMX debugging UI` didn't turn up this program, so it's more
than just one quick google search away.