Remove hardcoding of `ARCHFLAGS` for macOS machines
On M1 Silicon, this is *not* guaranteed to be `x86_64`.

Even more confusingly, it's not guaranteed to be `amr64` either... it
depends whether the code is compiled under the `Rosetta` translation
layer or not.

So for now, it's easier to remove this hardcoding altogether rather than
trying to be clever about it.

Background:
* https://stackoverflow.com/questions/64908116/correct-archflags-value-on-apple-silicon
* https://gitlab.kitware.com/cmake/cmake/-/issues/20989
* https://gitlab.kitware.com/cmake/cmake/-/issues/21554