Add new `/opt/homebrew/{bin,sbin}` to `PATH`
On apple silicon, homebrew now defaults to installing to
`/opt/homebrew/{bin,sbin}` instead of `/usr/local/{bin,sbin}`.

So add that to the path. Ideally I could add only one or the other
depending on the underlying silicon, but that starts to get overly
clever/complex, so simplest to just add both for now.

Note: Do not add `/opt/homebrew/{bin,sbin}` directly, as the braces
don't get expanded properly. That's why it's fully written out.

This needs to be added very early in the loading cycle because I set
several env vars using `$(brew --prefix)/bin/python` for example to
auto-detect which brew path to use.

Ideally this is something that prezto starts handling by default, as I
expect it will affect more and more users, so I filed
https://github.com/sorin-ionescu/prezto/issues/1988.