Switch VS Code to using its built-in settings sync
VS Code now offers a built-in settings sync, so switch to that instead
of managing through dotfiles.

It's very convenient that way, but doesn't give me a built-in versioned
history... so maybe down the road I'll switch back, but for now this is
probably more pragmatic. Especially since individual projects can
include their own workspace-specific overrides that can be committed to
the project.