Stop syncing python linters
It seems most projects are moving to `ruff`, including a couple of folks
I highly trust.

I no longer work primarily in Python, so haven't had the time to play
with it yet, but hope to soon.

In the meantime, no point in syncing these other linter settings across
machines/dotfiles.